Major Updates to data/Plotting Code/Fixed Panel Load Error
