--- data/ASOS_MPING_data.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-source /linuxapps/bin/condalatest.sh
-python ASOS_MPING_data.py
