+++ data/ASOS_MPING_data/ASOS_MPING_data.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+source /linuxapps/bin/condalatest.sh
+python ASOS_MPING_data.py
